major refactoring of ui tests, where actions ui tests like clicking a button are moved into TaskManagerController tests
