+++ src/test/java/edu/mikedev/task_manager/controller/TestUpdateTaskBindings.java
@@ -1,8 +1,9 @@
-package edu.mikedev.task_manager.ui;
+package edu.mikedev.task_manager.controller;
 
-import edu.mikedev.task_manager.model.Model;
 import edu.mikedev.task_manager.Task;
 import edu.mikedev.task_manager.User;
+import edu.mikedev.task_manager.model.Model;
+import edu.mikedev.task_manager.ui.LoginWindow;
 import edu.mikedev.task_manager.utils.UIScenarios;
 import org.assertj.swing.annotation.GUITest;
 import org.assertj.swing.core.matcher.JLabelMatcher;
@@ -15,43 +16,46 @@ import org.assertj.swing.util.Triple;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import javax.swing.*;
-import java.text.SimpleDateFormat;
 import java.util.List;
 
 import static edu.mikedev.task_manager.ui.UserTasksPage.htmlWrappedDescription;
 
 @RunWith(GUITestRunner.class)
-public class TestUpdateTask extends AssertJSwingJUnitTestCase{
+public class TestUpdateTaskBindings extends AssertJSwingJUnitTestCase{
 
-    JFrame window;
+    LoginWindow window;
     FrameFixture frame;
-    Task task;
+    List<Task> tasksListSorted;
+    TaskManagerController controller;
 
     @Override
     protected void onSetUp() {
         Triple<Model, User, List<Task>> scenario = UIScenarios.anyLoginUserTasksScenario();
-        task = scenario.third.get(0);
+        tasksListSorted = scenario.third;
         GuiActionRunner.execute(() ->{
-            window = new JFrame();
-            window.setContentPane(new NewUpdateTaskPage(task));
+            window = new LoginWindow();
             return window;
         });
+        controller = new TaskManagerController(scenario.first, window);
         frame = new FrameFixture(robot(), window);
         frame.show();
+
+        frame.button("btnLogin").click();
+        frame.panel("task0").click();
+        frame.button("btnUpdate").click();
     }
 
     @SuppressWarnings("java:S2699")
     @Test
     @GUITest
     public void testInitialState(){
+        frame.requireTitle("Update task \"Task 1\"");
         frame.label("lblTaskName").requireText("Task Name");
         frame.label("lblTaskDescription").requireText("Task Description");
         frame.label("lblTaskDeadline").requireText("Deadline (dd/MM/yyyy)");
-        frame.textBox("tfTaskName").requireText(task.getTitle());
-        frame.textBox("tfTaskDescription").requireText(task.getDescription());
-        SimpleDateFormat dateFormatter = new SimpleDateFormat("dd/MM/yyyy");
-        frame.textBox("tfTaskDeadline").requireText(dateFormatter.format(task.getDeadline()));
+        frame.textBox("tfTaskName").requireText("Task 1");
+        frame.textBox("tfTaskDescription").requireText("Description task 1");
+        frame.textBox("tfTaskDeadline").requireText("11/02/2014");
         frame.button("btnSave").requireText("Update");
     }
 }
