+++ src/test/java/edu/mikedev/task_manager/ui/TestLoginWindow.java
@@ -30,15 +30,8 @@ public class TestLoginWindow extends AssertJSwingJUnitTestCase {
 
 	@Override
 	protected void onSetUp(){
-		Model model = mock(Model.class);
-		when(model.areCredentialCorrect(ArgumentMatchers.matches("myusername"),
-				                        ArgumentMatchers.matches("mypassword"))).thenReturn(true);
-		User dummyUser = new User("myusername", "mypassword", "email@email.com");
-		Set<Task> taskSet = new HashSet<Task>();
-		dummyUser.setTasks(taskSet);
-		when(model.loginUser(anyString(), anyString())).thenReturn(dummyUser);
 		GuiActionRunner.execute(() ->{
-			window = new LoginWindow(model);
+			window = new LoginWindow();
 			return window;
 		});
 		frame = new FrameFixture(robot(), window);
@@ -56,52 +49,6 @@ public class TestLoginWindow extends AssertJSwingJUnitTestCase {
 		frame.label(JLabelMatcher.withName("lblErrorMessage")).requireNotVisible();
 	}
 
-    @SuppressWarnings("java:S2699")
-    @Test
-	@GUITest
-	public void testClickLoginAndThenErrorMessage(){
-		frame.label(JLabelMatcher.withName("lblErrorMessage")).requireNotVisible();  //to find not visible elements they must be found through Matcher class
-		frame.button("btnLogin").click();
-		frame.label("lblErrorMessage").requireVisible();
-		frame.label("lblErrorMessage").requireText("Username and/or password are wrong");
-	}
-
-    @SuppressWarnings("java:S2699")
-    @Test
-	@GUITest
-	public void testCorrectLogin() throws InterruptedException {
-		frame.textBox("tfUsername").click().enterText("myusername");
-		frame.textBox("tfPassword").click().enterText("mypassword");
-		frame.button("btnLogin").click();
-		frame.requireTitle("myusername tasks");
-	}
-
-    @SuppressWarnings("java:S2699")
-    @Test
-	@GUITest
-	public void testWrongPromptedLogin(){
-		frame.textBox("tfUsername").click().enterText("wrongusername");
-		frame.textBox("tfPassword").click().enterText("wrongpassword");
-		frame.button("btnLogin").click();
-
-		frame.requireTitle("Login page");
-
-		frame.label("lblUsername").requireText("Username");
-		frame.label("lblErrorMessage").requireEnabled();
-	}
-
-    @SuppressWarnings("java:S2699")
-    @Test
-	@GUITest
-	public void goToRegisterPage(){
-		frame.button("btnRegister").click();
-
-		frame.requireTitle("Registration page");
-		frame.label("lblUsername").requireEnabled();
-		frame.label("lblPassword").requireEnabled();
-		frame.label("lblEmail").requireEnabled();
-		frame.button("btnConfirmRegister").requireEnabled();
-	}
 
 
 	@After
