+++ src/test/java/edu/mikedev/task_manager/ui/TestRegistrationWindow.java
@@ -14,13 +14,15 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentMatchers;
 
+import javax.swing.*;
+
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @RunWith(GUITestRunner.class)
 public class TestRegistrationWindow extends AssertJSwingJUnitTestCase {
 
-    LoginWindow window;
+    JFrame window;
     FrameFixture frame;
 
     @Override
@@ -29,7 +31,8 @@ public class TestRegistrationWindow extends AssertJSwingJUnitTestCase {
         when(model.userExists(ArgumentMatchers.matches("existinguser1"))).thenReturn(true);
 
         GuiActionRunner.execute(() ->{
-            window = new LoginWindow(model);
+            window = new JFrame();
+            window.setContentPane(new RegistrationPage());
             return window;
         });
         frame = new FrameFixture(robot(), window);
@@ -48,65 +51,6 @@ public class TestRegistrationWindow extends AssertJSwingJUnitTestCase {
         lblErrorMessageUsername.requireNotVisible();
         lblErrorMessagePassword.requireNotVisible();
         lblErrorMessageEmail.requireNotVisible();
-        frame.requireTitle("Registration page");
-
-    }
-
-    @SuppressWarnings("java:S2699")
-    @Test
-    @GUITest
-    public void testEmptyFieldsRegistration(){
-        JLabelFixture lblErrorMessageUsername = frame.label(JLabelMatcher.withName("lblErrorMessageUsername"));
-        JLabelFixture lblErrorMessagePassword = frame.label(JLabelMatcher.withName("lblErrorMessagePassword"));
-        JLabelFixture lblErrorMessageEmail = frame.label(JLabelMatcher.withName("lblErrorMessageEmail"));
-
-        frame.button("btnConfirmRegister").click();
-
-        lblErrorMessageUsername.requireVisible();
-        lblErrorMessageUsername.requireText("Missing username");
-        lblErrorMessagePassword.requireVisible();
-        lblErrorMessagePassword.requireText("Missing password");
-        lblErrorMessageEmail.requireVisible();
-        lblErrorMessageEmail.requireText("Missing e-mail");
-    }
-
-    @SuppressWarnings("java:S2699")
-    @Test
-    @GUITest
-    public void testWrongFormattedEmail(){
-        frame.textBox("tfEmail").setText("This is not an email");
-
-        frame.button("btnConfirmRegister").click();
-
-        JLabelFixture lblErrorMessageEmail = frame.label(JLabelMatcher.withName("lblErrorMessageEmail"));
-        lblErrorMessageEmail.requireVisible();
-        lblErrorMessageEmail.requireText("The input prompted above is not an e-mail");
-    }
-
-    @SuppressWarnings("java:S2699")
-    @Test
-    @GUITest
-    public void testExistingUser(){
-        frame.textBox("tfUsername").setText("existinguser1");
-
-        frame.button("btnConfirmRegister").click();
-
-        JLabelFixture lblErrorMessageUsername = frame.label(JLabelMatcher.withName("lblErrorMessageUsername"));
-        lblErrorMessageUsername.requireVisible();
-        lblErrorMessageUsername.requireText("Username exists");
-    }
-
-    @SuppressWarnings("java:S2699")
-    @Test
-    @GUITest
-    public void testCorrectRegistration(){
-        frame.textBox("tfUsername").setText("newuser1");
-        frame.textBox("tfPassword").setText("password1");
-        frame.textBox("tfEmail").setText("email@email.it");
-
-        frame.button("btnConfirmRegister").click();
-
-        frame.dialog(DialogMatcher.withTitle("Registration completed")).requireVisible();
     }
